+++ Hi-Efficiency Bar/Hi-Efficiency Bar/Helpers/Objects/IngredientCusObj.swift
@@ -10,14 +10,14 @@ import UIKit
 
 class IngredientCusObj: NSObject {
      var id: Int?
-    var ratio: Int?
+    var ratio: Double?
     var name: String?
-    var unit: Int?
+    var unit: Double?
     var unit_view: String?
     var value: Double?
     init(dict: NSDictionary) {
-        self.ratio = dict["ratio"] as? Int
-        self.unit = dict["unit"] as? Int
+        self.ratio = dict["ratio"] as? Double
+        self.unit = dict["unit"] as? Double
         self.unit_view = dict["unit_view"] as? String
         if let val = dict["ingredient"] as? NSDictionary
         {
